refactor: decompose renderer workflows and harden electron bridge fallbacks
