+++ tests/unit/renderer/dashboard-storage.dom.test.js
@@ -89,6 +89,26 @@ describe('dashboard storage', () => {
     await expect(storage.hydrateConnectionSecret()).rejects.toThrow('missing secret');
   });
 
+  test('persistConnectionConfig usa fallback de sessionStorage sin Electron', async () => {
+    const originalElectronApi = window.electronApi;
+    delete window.electronApi;
+
+    try {
+      await storage.persistConnectionConfig({
+        provider: 'github',
+        organization: 'acme',
+        project: 'repo-a',
+        repositoryId: 'repo-a',
+        personalAccessToken: 'browser-secret',
+        targetReviewer: '',
+      });
+
+      await expect(storage.hydrateConnectionSecret()).resolves.toBe('browser-secret');
+    } finally {
+      window.electronApi = originalElectronApi;
+    }
+  });
+
   test('saved azure contexts sigue deshabilitado y limpia legacy', () => {
     window.localStorage.setItem(storage.DASHBOARD_SAVED_CONTEXTS_KEY, 'legacy-contexts');
